requirements.txt
Django>=4.0
djangorestframework>=3.13.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.3
